Fixed size display on install confirmation prompt Fixed wrong calculation when using invertModifier Adjusted install and unlock prices based on hull value Augments are in high demand and are now getting removed from market after you leave Added new quality filter Added CC support for adding augments Added 24 new augments

+++ src/de/schafunschaf/voidtec/util/MathUtils.java
@@ -17,14 +17,14 @@ public class MathUtils {
         return (int) (Math.round(number / pow) * pow);
     }
 
-    public static float roundWholeNumber(float number, int numPlaces) {
+    public static int roundWholeNumber(float number, int numPlaces) {
         double pow = Math.pow(10, numPlaces);
-        return (float) (Math.round(number / pow) * pow);
+        return (int) (Math.round(number / pow) * pow);
     }
 
-    public static double roundWholeNumber(double number, int numPlaces) {
+    public static int roundWholeNumber(double number, int numPlaces) {
         double pow = Math.pow(10, numPlaces);
-        return (Math.round(number / pow) * pow);
+        return (int) (Math.round(number / pow) * pow);
     }
 
     public static float roundDecimals(float number, int numDecimals) {
